+++ CobraBay/config/__init__.py
@@ -1,3 +1,10 @@
 """
 CobraBay - Configuration Handling
 """
+
+from .configs import CBCoreConfig, CBSMConfig
+
+all = [
+    'CBCoreConfig',
+    'CBSMConfig'
+]
